Setup database migrations
In merkle_tree/prisma, we put a migration `20230418081122_` that adds the tables necessary to save the Merkle trees.

In frontend/prisma, we put migrations `20230418081122_` and `20230418081148_add_api_tables` that add the tables for both Merkle trees and handling API requests.

This separation will allow us to add more tables for the API without affecting the code in merkle_tree/.